Cleaning, organizational, and documentation related changes.
