+++ PlayRho/Dynamics/WorldMisc.hpp
@@ -26,81 +26,15 @@
 /// Declarations of free functions of World for unidentified information.
 
 #include <PlayRho/Common/Math.hpp>
-#include <PlayRho/Common/Range.hpp> // for SizedRange
 
-#include <PlayRho/Collision/MassData.hpp>
-
-#include <PlayRho/Dynamics/BodyID.hpp>
-#include <PlayRho/Dynamics/FixtureID.hpp>
-#include <PlayRho/Dynamics/BodyConf.hpp> // for GetDefaultBodyConf
 #include <PlayRho/Dynamics/StepConf.hpp>
 #include <PlayRho/Dynamics/StepStats.hpp>
-#include <PlayRho/Dynamics/Contacts/KeyedContactID.hpp> // for KeyedContactPtr
-#include <PlayRho/Dynamics/FixtureConf.hpp>
-#include <PlayRho/Dynamics/WorldConf.hpp>
-#include <PlayRho/Dynamics/Joints/JointID.hpp>
-
-#include <iterator>
-#include <vector> // for std::vector and std::size
-#include <memory>
-#include <functional>
 
 namespace playrho {
-
-struct FixtureProxy;
-
 namespace d2 {
 
 class World;
-struct JointConf;
-class Manifold;
-class ContactImpulsesList;
 class DynamicTree;
-
-/// @brief Gets the bodies of the specified world.
-/// @relatedalso World
-SizedRange<std::vector<BodyID>::const_iterator> GetBodies(const World& world);
-
-/// @brief Gets the joints of the specified world.
-/// @relatedalso World
-SizedRange<std::vector<JointID>::const_iterator> GetJoints(const World& world);
-
-/// @brief Gets the contacts of the specified world.
-/// @relatedalso World
-SizedRange<std::vector<KeyedContactPtr>::const_iterator> GetContacts(const World& world);
-
-/// @brief Gets the body count in the given world.
-/// @return 0 or higher.
-/// @relatedalso World
-inline BodyCounter GetBodyCount(const World& world) noexcept
-{
-    using std::size;
-    return static_cast<BodyCounter>(size(GetBodies(world)));
-}
-
-/// Gets the count of joints in the given world.
-/// @return 0 or higher.
-/// @relatedalso World
-inline JointCounter GetJointCount(const World& world) noexcept
-{
-    using std::size;
-    return static_cast<JointCounter>(size(GetJoints(world)));
-}
-
-/// @brief Gets the count of contacts in the given world.
-/// @note Not all contacts are for shapes that are actually touching. Some contacts are for
-///   shapes which merely have overlapping AABBs.
-/// @return 0 or higher.
-/// @relatedalso World
-inline ContactCounter GetContactCount(const World& world) noexcept
-{
-    using std::size;
-    return static_cast<ContactCounter>(size(GetContacts(world)));
-}
-
-/// @brief Gets the touching count for the given world.
-/// @relatedalso World
-ContactCounter GetTouchingCount(const World& world) noexcept;
 
 /// @brief Steps the given world the specified amount.
 /// @relatedalso World
@@ -145,64 +79,9 @@ StepStats Step(World& world, Time delta,
 /// @relatedalso World
 const DynamicTree& GetTree(const World& world) noexcept;
 
-/// @copydoc World::GetFixturesForProxies
-/// @relatedalso World
-SizedRange<std::vector<FixtureID>::const_iterator>
-GetFixturesForProxies(const World& world) noexcept;
-
-/// @brief Gets the count of fixtures in the given world.
-/// @relatedalso World
-FixtureCounter GetFixtureCount(const World& world) noexcept;
-
 /// @brief Gets the count of unique shapes in the given world.
 /// @relatedalso World
-std::size_t GetShapeCount(const World& world) noexcept;
-
-/// @brief Gets the count of awake bodies in the given world.
-/// @relatedalso World
-BodyCounter GetAwakeCount(const World& world) noexcept;
-
-/// @brief Awakens all of the bodies in the given world.
-/// @details Calls all of the world's bodies' <code>SetAwake</code> method.
-/// @return Sum total of calls to bodies' <code>SetAwake</code> method that returned true.
-/// @see Body::SetAwake.
-/// @relatedalso World
-BodyCounter Awaken(World& world) noexcept;
-
-/// @brief Sets the accelerations of all the world's bodies.
-/// @param world World instance to set the acceleration of all contained bodies for.
-/// @param fn Function or functor with a signature like:
-///   <code>Acceleration (*fn)(const Body& body)</code>.
-/// @relatedalso World
-template <class F>
-void SetAccelerations(World& world, F fn)
-{
-    const auto bodies = GetBodies(world);
-    std::for_each(begin(bodies), end(bodies), [&](const auto &b) {
-        SetAcceleration(world, b, fn(world, b));
-    });
-}
-
-/// @brief Sets the accelerations of all the world's bodies to the given value.
-/// @relatedalso World
-void SetAccelerations(World& world, Acceleration acceleration) noexcept;
-
-/// @brief Sets the accelerations of all the world's bodies to the given value.
-/// @note This will leave the angular acceleration alone.
-/// @relatedalso World
-void SetAccelerations(World& world, LinearAcceleration2 acceleration) noexcept;
-
-/// @brief Clears forces.
-/// @details Manually clear the force buffer on all bodies.
-/// @relatedalso World
-inline void ClearForces(World& world) noexcept
-{
-    SetAccelerations(world, Acceleration{});
-}
-
-/// @brief Finds body in given world that's closest to the given location.
-/// @relatedalso World
-BodyID FindClosestBody(const World& world, Length2 location) noexcept;
+FixtureCounter GetShapeCount(const World& world) noexcept;
 
 /// @brief Gets the min vertex radius that shapes for the given world are allowed to be.
 /// @relatedalso World
@@ -211,10 +90,6 @@ Length GetMinVertexRadius(const World& world) noexcept;
 /// @brief Gets the max vertex radius that shapes for the given world are allowed to be.
 /// @relatedalso World
 Length GetMaxVertexRadius(const World& world) noexcept;
-
-/// @brief Gets the bodies-for-proxies range for the given world.
-SizedRange<std::vector<BodyID>::const_iterator>
-GetBodiesForProxies(const World& world) noexcept;
 
 } // namespace d2
 } // namespace playrho
